refactor: create component UserPhoto
